refactor: ランダム起動ボタンを AppHeader からソートドロップダウン横に移動
再読込ボタンとの誤タップを防止するため、ソートと同じ文脈（toolbar 領域）に配置。
